tast: Add basic ARC support to local/chrome.
Add an ARCEnabled option that can be passed to chrome.New to
disable ARC opt-in verification, enable the Play Store via
the autotestPrivate API, and wait for Android to report that
it's finished booting.

Also add a trivial ui.ARCSanity local test that exercises
this code.

BUG=chromium:737628
TEST=ui.ARCSanity passes

Change-Id: Iec4092d0f478f59d41f059e1b10b6942ab7aa9c1
Reviewed-on: https://chromium-review.googlesource.com/671678
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>