tast-tests: Add NoLogin option to local/chrome.
Add a new NoLogin option that can be passed to
chrome.NewChrome to avoid automatically logging in. This can
be used by tests that want to verify behavior in a
not-logged in state but still need to interact with Chrome
(and hence want to restart it with testing-friendly flags).

BUG=chromium:770562
TEST=existing ui.ChromeSanity test passes, and used it in
     a new test and verified that it puts chrome into the
     expected state

Change-Id: I0da0daecb5b07be1f8747dee6acdee3d14f12f12
Reviewed-on: https://chromium-review.googlesource.com/826163
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>