tast-tests: Rename ui.ChromeSanity and ui.ARCSanity tests.
Rename the ui.ChromeSanity test to ui.ChromeLogin to more
accurately reflect what it does and to make it clearer that
it requires the ability to log in (which isn't supported on
some systems).

Also rename ui.ARCSanity to ui.ARCStartup.

BUG=none
TEST=ran tests matched by "(bvt)" expression and verified
     that renamed tests passed

Change-Id: I2f90bfa18b75faecd102559368c61603fc085d89
Reviewed-on: https://chromium-review.googlesource.com/1020143
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>