tast-tests: Minor improvements to chrome package.
Introduce a new chrome.Chrome.NewConnForTarget method that
returns a connection to the first target matched by a
caller-provided function.

Also update chrome.Conn.EvalPromise to check for an
exception before trying to unmarshal the resulting value,
and improve the internal poll() function to return the
last-seen error instead of the generic "context deadline
exceeded" error produced by the context.

Finally, make minor improvements and corrections to some
comments for the Conn type.

BUG=none
TEST=manually verified that a nicer error message is
     returned when an exception is encountered during
     EvalPromise; also verified that existing UI tests pass

Change-Id: Ibd1c257d3bfaf3d47073e5d898b5feab12ecf699