Update patch set 2
Patch Set 2:

We probably need the video team's input. I assume the bitstream needs a minimum amount of complexity for the codec, so pure color is probably not a good thing.

That said, yes we want the binaries on the test image, if only to reduce network issues leading to test failures. We could either have an extra git repo for tast that only contains binaries (and purge this periodically?), or an ebuild that copies a list of binaries from a GS location to the test images (ideally only one ebuild for all of tast, not one per test as we do right now for autotest).

Patch-set: 2
CC: Gerrit User 1196199 <1196199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1001849 <1001849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1002450 <1002450@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>