Update patch set 3
Patch Set 3:

(5 comments)

My concerns are primarily to leave in the error case enough breadcrumbs in the logs to show what the string were leading to api parsing failures. They don't have to be spewed all the time, but if there is a problem it would be good to see the problem quickly.

Patch-set: 3
Reviewer: Gerrit User 1000221 <1000221@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>