Port the GraphicsApiHelper functionality to tast.
This CL ports the functionality of the GraphicsApiHelper class defined
in autotest/files/client/cros/graphics/graphics_utils.py to the tast
graphics support package.

The main functions that calling code should use are SupportedAPIs() and
DEQPExecutable(). However, GLESVersion() and SupportsVulkanForDEQP() are
also exported so that the calling code can get the necesary arguments to
pass to the main functions.

TEST=added unit tests to this CL
BUG=chromium:865107
CQ-DEPEND=CL:1144238

Change-Id: I3826ec42bc70b11d40eee7c242e0238c130ba5de