Add the ParseDEQPOutput function to the graphics support package.
This CL adds the ParseDEQPOutput function to the graphics support
package to parse log files generated by DEQP tests. The parsing is
essentially a state machine, so for ease of review, please refer to the
diagram at https://tinyurl.com/deqp-parsing-fsm.

The DEQPOutcomeIsFailure function is also added. It's exported in case
callers need to interpret the keys of the |stats| map returned by the
ParseDEQPOutput.

TEST=added unit tests to this CL
BUG=chromium:865107
CQ-DEPEND=CL:1153633

Change-Id: I3c85209724f92ca2c2890dc0cbeb11dcc5bb869f