tast-tests: Increase timeout for vm.CrostiniFiles test
This increases the overall timeout from 5 to 10 minutes. This is needed
because the test can take in the 5-10 minute range to complete on some
of the braswell boards. This test also uses the default qcow2 image
which is much slower since it goes through the Chrome UI to actually run
the test. The other VM tests which talk to concierge directly over D-Bus
use raw images which go much faster.

BUG=None
TEST=Verified test still works on eve

Change-Id: I892d94b407245376be799bf5bdc5516bb7801169
Reviewed-on: https://chromium-review.googlesource.com/1195949
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>