tast-tests: Add platform.Histograms test.
Add a new chromiumos/tast/local/chrome/metrics package
containing a GetHistogram function that can be used to get
histograms from Chrome. It also contains WaitForHistogram
and WaitForHistogramUpdate helper functions that can be used
to wait for a histogram to appear or for a histogram to be
updated.

Also add a new platform.Histograms test that calls
metrics_client to verify the libmetrics->Chrome path that
Chrome OS system daemons use to report metrics.

BUG=chromium:878641
TEST=platform.Histograms and new unit tests pass

Change-Id: I71604c347314bed4730b6bfcdfe9b0ecc24de7db