tast-tests: Make "Crosvm" naming consistent.
Rename instances of "CrosVM" to "Crosvm" to follow the
existing convention of using underscores to separate words
in filenames and capitalization to separate words in
function (and test) names.

The naming of the vm.StartCrosVM test conflicts with a
change I'm working on to verify that filenames and test
names are aligned, since it's registered by a file named
start_crosvm.go (rather than start_cros_vm.go).

The vm.CrosVM struct currently lives in cros_vm.go, so
another option would be to rename the test file to
start_cros_vm.go. As I understand it, the actual hypervisor
that's being tested is named "crosvm", though, so it feels a
bit less surprising to me to make the filenames follow that
and change the capitalization in code to match.

(I also see that the name was written as "Crosvm" in the
subject of the initial announcement at
http://g/chromeos-rust/XNVRa6oy3-0/pCVc5_qIBAAJ.)

BUG=chromium:880661
TEST=updated "tast run" doesn't complain about test name

Change-Id: I49c68dec2cc0a5edbf0743a1e0781d341c6a509d
Reviewed-on: https://chromium-review.googlesource.com/1207212
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>