tast-tests: screenshot: Extract screenshot check into its own package
This extracts the functionality that takes a screenshot, creates a
histogram of the pixels and then checks if the dominant color matches a
target color and covers a certain ratio of the screen. This is only used
by the Screenshot test right now, but new tests will be added soon that
also use this functionality.

BUG=chromium:849438
TEST=tast run eve graphics.Screenshot

Change-Id: Ic39282305f9673fe0266c9b8156d9e8578635a5f