tast-tests: Test Linux package info and installation for Crostini
This copies a .deb file that we created for testing purposes to the
container and then executes the cicerone D-Bus call for getting the
information on the package file and verifies the result. It then
executes another D-Bus call to cicerone which actually does an install
on the package file and verifies that result and that the files were
actually installed into the container.

Additional tests will be built on top of this that verify the .desktop
files the .deb package installs show up in the launcher, have icons and
work as expected.

The .deb package file is built from the source here:
https://chromium.googlesource.com/chromiumos/containers/cros-container-guest-tools/+/master/cros-tast-tests/

BUG=chromium:849439
TEST=tast run eve vm.CrostiniStartEverything

Change-Id: I57452089e7aaef4fa392af19bfbf5f8f197de856
Reviewed-on: https://chromium-review.googlesource.com/1225111
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>