tast-tests: Add launcher integration tests for Crostini
This adds tests which verify that the package installed in prior CLs
actually ends up with the items integrated into the Chrome launcher,
that their icons are in the filesystem, the apps can be launched, they
draw to the screen, their shelf items are present and then in a later CL
we will close them and verify their shelf items disappear.

NOTE: The screenshot tests from the launcher currently fail for the x11
app due to https://crbug.com/887016.

BUG=chromium:849438
TEST=tast run vm.CrostiniStartEverything
CQ-DEPEND=CL:1235199

Change-Id: Ic65c1cf3ed4f0bbb115c5a9bc425ca842eb16a7c
Reviewed-on: https://chromium-review.googlesource.com/1234304
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>