Update patch set 5
Patch Set 5:

> Patch Set 5: Verified+1
> 
> (2 comments)
> 
> > Patch Set 4: Code-Review+1
> > 
> > (2 comments)
> > 
> > lgtm with minor suggestions; feel free to self +2 after addressing
> 
> Thanks Dan

By the way, I have set this to "informational" as recommended in the docs, which I understand means it doesn't block stuff if it fails. Which it will, because it depends on USE=ml_service, which is not set by default.

When I do set USE=ml_service by default soon, it might not be for every board ... what's the right way to handle that if this test becomes non-"informational"?

Patch-set: 5
Reviewer: Gerrit User 1182889 <1182889@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>