Add new ui.VirtualkeyboardTyping test.
Add a new test which checks that tapping keys on the virtual
keyboard sends those keys to a text field on a web page.

Had to use the virtualKeyboardPrivate chrome API to bring up
the VK. virtualKeyboardPrivate is only allowed on whitelisted
extensions, so we had to add another chrome flag.

BUG=chromium:879073

TEST=tested with eve device.

Change-Id: I21ef9232dc81f0953913acc3bd352868e4411c19