Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5: Verified+1
> > 
> > (2 comments)
> > 
> > > Patch Set 4: Code-Review+1
> > > 
> > > (2 comments)
> > > 
> > > lgtm with minor suggestions; feel free to self +2 after addressing
> > 
> > Thanks Dan
> 
> By the way, I have set this to "informational" as recommended in the docs, which I understand means it doesn't block stuff if it fails. Which it will, because it depends on USE=ml_service, which is not set by default.
> 
> When I do set USE=ml_service by default soon, it might not be for every board ... what's the right way to handle that if this test becomes non-"informational"?

oh, i see. yeah, please don't add this in its present form if it's guaranteed to fail on some of the boards where it'll be running. that'll result in me getting spammed, at the very least. :-)

the right way to handle this is to introduce a new "ml_service" software feature, as described in go/tast-deps:

- add ml_service to IUSE in the tast-use-flags ebuild
- update https://chromium.googlesource.com/chromiumos/platform/tast/+/master/src/chromiumos/cmd/local_test_runner/main.go to define a new ml_service feature that's tied to the USE flag
- list the ml_service feature name in the SoftwareDeps field of the test

then the test will be automatically skipped when tast is run on systems that weren't built with the ml_service USE flag.

Patch-set: 5
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>