Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (2 comments)
> > 
> > Nick,
> > 
> > Do you know of some way I can check that the uninstall succeeded/failed without having to use the histogram?  I could just invoke CrostiniManager::RemoveCrostini myself as an alternative but figured it'd be better to test this through the actual UI.
> 
> I think a good option would be for CrostiniManager to allow observers to register for the uninstall action.  Then in the AutotestPrivateRunCrostiniUninstallerFunction you could register for that event, uninstall via UI, then wait for notification before invoking JS callback.  I can do a CL right now to add that observer.

That would definitely work...it's almost the end of my day so if you want to go ahead and do that maybe it'll be merged by the time I'm in tomorrow and then I can hook it up to this. :) I'm assuming it would indicate success of failure so the test will know.

Patch-set: 3
Reviewer: Gerrit User 1298800 <1298800@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>