tast-tests: Add short/full versions of WebRTC tests
Shorten the performance time of existing video.WebRTC tests and
add three video.WebRTC*Perf tests as the full versions.
While short versions are intended to be run in Chrome CQ, where tests
have to be done in several seconds, the full versions will be used
for gathering the performance data in waterfall.

This commit also moves the starting points for counting time, i.e.
the places where setTimeout are called in getusermedia.html and
loopback.html. This is because we want to exclude the time for
completing getUserMedia calls.

BUG=chromium:882433, chromium:852302
TEST=Run each test on kevin by 'tast run'
TEST=Run each test on amd64-generic VM by 'tast run'

Change-Id: I1852b86c23e53beb271cb9ff6d4c9f09f0285f62
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1237804
Reviewed-by: Shuhei Takahashi <nya@chromium.org>