tast-tests: Remove vm.StopConcierge call
I always comment this out during development because it's very helpful
to leave the VM/container up for introspection after a test fails. Since
we aren't actually checking the result of it, we are not actually
testing anything with this call. Changes I've made to the CrostiniFiles
test now do uninstall which will verify we can stop the VM successfully,
so we do have coverage of that aspect in that test which should make it
safe to remove from here.

BUG=chromium:849438
TEST=tast run vm.CrostiniStartEverything

Change-Id: I216f601e79b077dd4cc7103d8324adf42a3052ee
Reviewed-on: https://chromium-review.googlesource.com/1239195
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>