tast-tests: Add input constants and generator script.
Add a short script for generating constants from the
kernel's include/uapi/linux/input-event-codes.h header,
along with the generated constants themselves. This is the
first part of a forthcoming input package.

The constants were generated using the following commands:

  go run gen/gen_constants.go \
    ../../../../../../../third_party/kernel/v4.14/include/uapi/linux/input-event-codes.h \
    generated_constants.go

  go fmt generated_constants.go

After a later change, running "go generate" will execute
these commands. This only needs to be done if/when the
kernel header changes.

BUG=chromium:878734
TEST=manual inspection (and verified in conjunction with
     later changes)

Change-Id: Id72e236ce9b6dafada3177e3a7836f04805060d0