Update patch set 4
Patch Set 4:

(3 comments)

derat@, ihf@: I think we should probably agree on the sync issue. It seems that if there is a lot of pending data (maybe non-logs) and we don't do a full sync, there will still be heavy I/O work later on because of the modification to the dirty_writeback_centisecs setting and this might delay the flushing of logs during the test. I guess this is what Ilja meant by competing writes. A full sync might prevent the test from starting quick in certain situations, but at least the logs might retain more information. So, for the sake of consistency with the current autotest version, would it be so bad at this point to do the full sync and refine later if there is a performance problem?

Patch-set: 4
Reviewer: Gerrit User 1273510 <1273510@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>