Update patch set 4
Patch Set 4: Verified-1

(build sheriff here)

Could this CL be related to the failure of the audio_Microphone autotest in CQ?

https://cros-goldeneye.corp.google.com/chromeos/healthmonitoring/buildDetails?buildbucketId=8934351856400623232

> 09/26 00:20:35.093 DEBUG|  audio_Microphone:0051| Selected input device name=chtrt5650: :1,0
> 09/26 00:20:35.095 DEBUG|             utils:0219| Running 'alsa_helpers --device hw:1,0 --get_capture_channels'
> 09/26 00:20:35.106 ERROR|             utils:0287| [stderr] Assign device to hw:1,0
> 09/26 00:20:35.117 ERROR|             utils:0287| [stderr] Cannot open audio device for capture: hw:1,0 (Device or resource busy)
> 09/26 00:20:35.124 ERROR|        alsa_utils:0198| Fail to get supported channels for hw:1,0

Sorry if this CL is unrelated - and I do see that the failed test is a Python one while this CL is a Tast one. But could it happen that these two tests conquered for access to the same audio device?

Patch-set: 4
Reviewer: Gerrit User 1126309 <1126309@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=-1