tast-tests: Wait for ComponentUpdater for the StartCrosvm test
This test was failing because example.DBus was restarting the ui job as
part of its test, which caused the ComponentLoader service to not be
found when running this test. Now we wait for that to be started up
before running this test.

BUG=chromium:849438
TEST=tast run eve example.DBus vm.StartCrosvm

Change-Id: I626636fc3fa477b9f998abb56a642198ede80bd9
Reviewed-on: https://chromium-review.googlesource.com/1246276
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>