tast-tests: Startup Chrome for the StartCrosvm test
This test was failing because example.DBus was restarting the ui job as
part of its test, which caused the ComponentLoader service to not be
found when running this test. Since this test has a dependency on
something running in Chrome (although not on being logged in), we do
ensure Chrome is running as part of this test initially.

BUG=chromium:849438
TEST=tast run eve example.DBus vm.StartCrosvm

Change-Id: I626636fc3fa477b9f998abb56a642198ede80bd9