Update patch set 4
Patch Set 4:

I am fine with this change, but a full sync is the correct thing to do here. Not only is it for being nice to the tests that ran before, but also increases our chances to find something useful in the logs of the graphics test. dirty_writeback_centisecs will pretty much force the sync one second into the test, which of course is a second late and can cause a delay then.

Of course even with this there is no guarantee that we have something in the logs, but the quality of the logs went up a lot since we had this (or maybe we don't have so bad crashes anymore?)

Anyhow, I do recommend applying the dirty_writeback_centisecs to all of tast, not just graphics. Syncing logs to disk only once every 10 minutes is shitty.

Patch-set: 4
Reviewer: Gerrit User 1000221 <1000221@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>