Make vkb tast tests less flaky.
Recently we tried to remove the "informational" attribute from
ui.VirtualKeyboardOmnibox, but it caused some flaky failures.
The source of the flakiness might be from our use of the automation
API: the 'loadComplete' event only fires once, but we listen for
it multiple times, so the other times may not fire.

To fix this, we listen for 'loadComplete' once at the beginning
of the test and cache the desktop root node. We then simply
re-use this root node instead of calling getDesktop again. We
add a JavaScript helper function to do this.

BUG=chromium:888884,chromium:879073
TEST=tested on eve

Change-Id: I4d07370406f6a16b677554b49fefa8805a2fb5f0