tast-tests: Fix localhost prompt checking in StartCrosvm
Previously this test worked because the VM was logging out a line from
the kernel after the bash prompt was shown, which then caused a line
terminator to come through stdout. There was a kernel change which made
that line come earlier, so we are not seeing the line terminator on
stdout after the 'localhost #' prompt like we used to. This fixes that
problem by scanning for bytes from stdout rather than lines.

BUG=chromium:878908
TEST=tast run vm.StartCrosvm

Change-Id: I3aebffe588a0c9a6ae9f012e9d74877cb1833455