Make virtual keyboard tast tests less flaky.
Recently we tried to remove the "informational" attribute from
ui.VirtualKeyboardOmnibox, but it caused some flaky failures.
The source of the flakiness might be from our use of the automation
API: the 'loadComplete' event only fires once, but we listen for
it multiple times, so the other times may not fire.

To fix this, we just repeatedly poll whenever we are waiting for
an element to appear. We add a new WaitForPromise helper function
that complements WaitForExpr, waiting until a promise resolves to
true. We had to replace |vkb.IsShown| with |vkb.WaitForHidden| as
the former requires returning a boolean value from a promise, which
is not possible in |WaitForPromise| (since returning false will
just keep waiting instead of returning).

We also considered a more "proper" solution of listening only
re-querying on changes to the tree, but A) this requires adding
a JavaScript helper function and B) there could be a lot of tree
changes which means the test could run quite slowly.

BUG=chromium:888884,chromium:879073
TEST=tested on eve

Change-Id: I0df98ad7ebd9d31d0089392e256951c77bc6cf39