tast-tests: Add ui restart hack to upstart.StopJob.
The ui Upstart job is atypical, in that it's respawned by a
separate ui-respawn job that is started on "stopping ui". As
a result, if session_manager exits just before
upstart.StopJob(ctx, "ui") is called, the job may get
restarted even after the test runs "initctl stop ui".

This change makes upstart.StopJob wait for the ui job to
either have a "start" goal or reach "stop/waiting" while the
ui-respawn job is in "stop-waiting".

BUG=chromium:891594
TEST=ran ui.SupervisedUserCrash followed by
     ui.VirtualKeyboardOmnibox and saw the second test start
     chrome instead of getting an error like "Failed to
     start Chrome: stopping "ui" failed: status
     start/post-stop has non-"stop" goal"

Change-Id: I2d542e3939cee6369fe477b2776cf5c39e51cd60