tast-tests: Add ui restart hack to upstart.StopJob.
The ui Upstart job is atypical, in that it's respawned by a
separate ui-respawn job that is started on "stopping ui". As
a result, if session_manager exits just before
upstart.StopJob(ctx, "ui") is called, the job may get
restarted even after the test runs "initctl stop ui".

This change makes upstart.StopJob run "initctl stop ui" a
second time in this case. When ui-respawn sees that the job
was stopped manually instead of in response to
session_manager exiting, it leaves it alone.

BUG=chromium:891594
TEST=ran ui.SupervisedUserCrash and sequence
     ui.VirtualKeyboardOmnibox and saw a message like "ui
     job has status start/post-stop (probably due to
     ui-respawn); stopping it again" instead of an error
     like "Failed to start Chrome: stopping "ui" failed:
     status start/post-stop has non-"stop" goal"

Change-Id: I2d542e3939cee6369fe477b2776cf5c39e51cd60