Update patch set 1
Patch Set 1:

(2 comments)

Thanks for looking into this issue. It's an interesting and challenging problem.

BTW this seems like a general problem that DUT is in a "busy" state when a test started. We may want to consider writing testSetupFunc to wait for the DUT to settle before starting tests. In this case, we might be able to mitigate the problem by checking:
- the ui job is in a "stable" state (start/running or stop/waiting)
- the ui-respawn job is in stop/waiting state

Patch-set: 1
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>