Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hi all,
>   I'm trying to merge changes into vm.CrostiniStartEverything but I couldn't pass the test on a clean master branch (just repo sync-ed):
> vm.CrostiniStartEverything  [ FAIL ] webserver: Expected page content "nothing but the web", got ""
>   Before I dive into details, is there any quick thought about what might be happening ? 
>   The full log is at https://paste.googleplex.com/5440944417538048
> 
> Thank you.

Are you sure you're really on the latest? That looks exactly like an old bug in vm/subtest/webserver.go where we didn't wait for the python webserver to fully startup before testing it. This was the CL that fixed it: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/1225370

Check to make sure you have that change in your version...if you don't, then you are not on the latest. (you may have repo sync'd...but maybe you forgot to repo rebase)

Patch-set: 5