Update patch set 4
Patch Set 4:

(1 comment)

derat@, ihf@ This is a minimal port of the DEQP in tests split in two CLs. This CL is initialization. The other one in the relation chain is the actual running of the binaries. The second one is very close to what I'd like to land, but there might be minor changes.

Patch-set: 4
Reviewer: Gerrit User 1273510 <1273510@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false