tast-tests: Add chrome.LoggedIn precondition.
Add a new precondition to the chrome package that tests can
list in their "Pre" field to indicate that they require a
logged-in Chrome session. This lets the same Chrome session
be shared across multiple tests, which has the potential to
greatly reduce the amount of time needed to run test suites.

BUG=chromium:892009
TEST=ran updated tests and saw combined running time of
     these tests drop from 60+ seconds to 25 seconds on
     caroline

Change-Id: I29d0a59fc7e3c99c8cc4901405663abe9a8eb0c4