tast-tests: Use QCOW2 format for CrostiniStartEverything
There's an issue in the 3.18 kernel config that causes the RAW disk
format being used in this test to be orders of magnitude slower during
formatting than the QCOW2 format. This essentially undoes the gains we
get by using RAW disk format on the boards that need it most. For now we
will switch to QCOW2 format until we have the kernel changed to be more
optimal. (Although there is another open bug to switch back to QCOW2
after we make that format more optimal interestingly enough).

BUG=chromium:894521,chromium:876482
TEST=tast run vm.CrostiniStartEverything

Change-Id: I77c896260dc90dd53df317c59781dbfeee10328e