tast-tests: Update errors to mention Chrome crashes.
When the chrome package returns errors, it also checks if
the browser process's PID changed and mentions it if so. I
don't think I've ever seen this happen without having been
caused by a Chrome crash, so include "Chrome probably
crashed" to point people in the right direction when they're
investigating test failures.

BUG=none
TEST=none

Change-Id: I89e0ff8416d4b481df403a9c3b5dd4b628110d02