tast-tests: Make process tests wait for system-services.
Update the platform.CheckProcesses and
security.SELinuxProcessContext tests to wait for
the system-services Upstart job to be running. This
hopefully prevents failures when tests run immediately after
the system boots, which can happen on VM builders.

Also make platform.CheckProcesses wait 10 seconds rather
than 5 for Upstart jobs, and update
security.SELinuxProcessContext to permit non-critical
processes which may get restarted by other tests to not be
running and to not log multiline messages (which make test
logs difficult to read).

BUG=chromium:897521
TEST=manual: ran tests

Change-Id: I6d4478571d7d1b714f0d31134675c8be7972b3ed