Update patch set 3
Patch Set 3:

(9 comments)

as a high-level comment, having a test that just watches a process for 30 seconds to make sure that it doesn't restart seems a bit odd (and slow). if shill is in a crash loop, do we stop respawning it eventually? the platform.CheckProcesses test already makes sure that shill is running. as i understand it, that test is already intended to catch crashing processes (under the assumption that we won't respawn them forever).

Patch-set: 3
CC: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>