Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (9 comments)
> 
> as a high-level comment, having a test that just watches a process for 30 seconds to make sure that it doesn't restart seems a bit odd (and slow). if shill is in a crash loop, do we stop respawning it eventually? the platform.CheckProcesses test already makes sure that shill is running. as i understand it, that test is already intended to catch crashing processes (under the assumption that we won't respawn them forever).

We don't stop respawning it as we need shill for any interface's connectivity (it starts dhcpcd for all interfaces). While that design itself is debatable (it predates most of us), until it is there, we do have a need to check for a constant and not-changing PID. 
Just checking that shill is running doesn't cover that case?

Patch-set: 3
Reviewer: Gerrit User 1158909 <1158909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>