tast-tests: Clear policy files from /var/lib/whitelist.
Update chrome.New to delete the /var/lib/whitelist directory
when restarting the ui job. This is apparently needed in
some cases to get the DUT into a usable state if an earlier
test lets session_manager regenerate the owner.key file and
then kills it before it can also regenerate policy.1 and
sign it using the new owner key.

This also makes us clear the user's cryptohome while the ui
job is stopped rather than while it's running. I don't think
it's necessary, but it feels cleaner to delete policy files
and the cryptohome at around the same time.

BUG=chromium:897278
TEST=ran ui.ChromeLoginForever for 30 iterations without
     seeing failures; also started tryjobs

Change-Id: I9cf5226b7d82bad9f6fb7845815964cc1c9e0b6f