video.EncodeAccelH264*: Add large test cases
This CL adds large test cases for video encoding. They are 1080p, 720p, 360p and
180p video raw data.
Since their size are too large, we are not put them on DUT by default. They are
compressed and the compressed videos are put in DUT. So the test decodes them at
first, uses the raw data in the test and remove them in the end.

BUG=chromium:889496
TEST=video.EncodeAccelH264*
CQ-DEPEND=CL:1316967

Change-Id: Iad57ebc9b2d74be8cee9cf598542c38d36186170