WIP: add platform.Memd tast-test
This test checks that memd is running and responding to
events of interest by creating clip files.

The events of interest are faked.  A tab discard is simulated
by sending a D-Bus message that looks as if it came from Chrome.
A kernel OOM kill is simulated by producing a syslog line that
is recognized by the anomaly_collector daemon.  Thus this also
tests the OOM-kill reporting functionality of the anomaly collector.

BUG=none
TEST=none

Change-Id: Ibd278e4afe346dacd3ac37f8b10ce5a62d00a445