add platform.Memd tast-test
This test checks that memd is running and responding to
events of interest by creating clip files.

The events of interest are faked.  A tab discard is simulated
by sending a D-Bus message that looks as if it came from Chrome.
A kernel OOM kill is simulated by producing a syslog line that
is recognized by the anomaly_collector daemon.  Thus this also
tests the OOM-kill reporting functionality of the anomaly collector.

BUG=chromium:900291
TEST=ran the test
CQ-DEPEND=CL:1308497,CL:1308673

Change-Id: Ibd278e4afe346dacd3ac37f8b10ce5a62d00a445
Reviewed-on: https://chromium-review.googlesource.com/1306913
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>