Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> There is a tast vmtest failure on betty-arc64-paladin: https://cros-goldeneye.corp.google.com/chromeos/healthmonitoring/buildDetails?buildbucketId=8931027719439336288
> 
> The 2rd attempt may be caused by infra failure. Can you please take a look if the first attempt looks suspicious to you?

sorry, i was OOO today and wasn't expecting this to go in after the earlier failure, but it got auto-CQ+1-ed. :-(

yes, there's a race on VM builders with this change that can lead to test failures. i've uploaded https://crrev.com/c/1314050, which i hope will fix the issue, but reverting this change is also an option.

Patch-set: 5