Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> On a related note: if I were to look at migrating other sorts of suites to run in tast (e.g., our Wifi functionality test suites), how would that translate into Tast / autotest metadata-like changes? I see the tast control files in autotest's server/site_tests/tast/*. And I guess I would still rely on autotest's 'DEPENDENCIES = "wificell"' to restrict to devices with available hardware.
> 
> But then, do I define a new 'group:' (Attr: []string{"group:wifi_matfunc"} or similar)? And what if I wanted the same test to run in both BVT and 'other' suites? It doesn't appear that's easy to do right now, since control.mustpass-system checks for !"group:*".
> 
> I ask because network_ShillStability is currently run in both bvt and wifi_matfunc. It may not really need to, since the other Wifi tests will surely catch us if shill is crashing. But eventually we may have others that fall in a similar category.

how does the "wificell" autotest dependency get assigned? does it correspond to a USE flag, or is it a runtime decision that depends on the configuration of a particular DUT in the lab? for example, could you have caroline DUTs both with and without this dependency depending on how they're set up?

if it's a runtime decision, that's something that's unfortunately out of tast's hands, since autotest still manages where tests get scheduled (since it, you know, manages the whole lab of DUTs :-P). to restrict your test to run on DUTs that have been configured in a certain way, you may need to introduce a new control file for the tast autotest test that lists the autotest dependency that you need.

mind filing a bug with some more detail so we can figure out the best way to support this? thanks!

Patch-set: 1
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>