tast-tests: Add test for multifaceted memory usage
memoryuser contains the basic structure of running ARC
apps, opening Chrome tabs, and running commands in a VM
in parallel. Also collects /proc/meminfo, iostat -c,
/var/log/vmlog/ and /var/log/memd/ information.

The core functionality of running ARC, Chrome, and VMs together and
collecting logs has been separated out into the memoryuser directory
to make it easier to replace the specific apps, tabs, and vm commands
as well as adding additional tests.

heavy_memory_user is a test that opens Chrome tabs and runs dd in a
vm in parallel to use a large amount of memory.

BUG=chromium:902164
TEST=Ran the test

Change-Id: Idc76ee6abc3387a00d537442d550049798b9463d