Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> > Patch Set 2:
> > 
> > (1 comment)
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > (1 comment)
> > > > 
> > > > Hi! Thanks for dealing with the bug.
> > > > 
> > > > According to the bug thread, getDesktop throws an exception, so I'm afraid this change can not catch it. I guess we need to call getDesktop inside try-catch clause and retry.
> > > 
> > > Yeah I wasn't sure what the stack trace meant:
> > > 
> > >         at find (extensions::automationNode:676:16)
> > >         at publicClassPrototype.(anonymous function) (extensions::utils:137:25)
> > >         at check (.:4:25)
> > >         at chrome.automation.getDesktop.root (.:11:2)
> > >         at  (extensions::automation:149:6)
> > > 
> > > I guess it could be a) getDesktop calling find, or b) we are calling root.find(). I was guessing the latter because it says getDesktop.root.
> > > 
> > > I also noticed that this error is happening in WaitUntilShown too, so I'll add the same code to that as well.
> > 
> > In fact, the error probably has something to do with calling getDesktop multiple times. Let's see how this fix goes first and then we can try just calling getDesktop once later.
> 
> Hmm, sorry, I was misreading the stack trace. It says chrome.automation.getDesktop() calls the callback function (check) synchronously, which I did not expect. :(
> 
> But since the error says there was an unhandled exception, I'm afraid just checking return value won't improve the situation. We somehow need to catch exceptions.

Oh, sorry for not being clear. I wasn't trying to catch the exception. I was trying to (speculatively) stop the exception from occurring, by polling on "loadComplete" event instead of polling at regular intervals. However, it probably won't fix anythng :/ It's really hard to find the root cause without a repro :(

I think try/catch should definitely fix the exception error, but not the flaky timeouts that are also happening sometimes. The issue is definitely with the automation API / how I'm using the API, because the screenshots in the logs show that the virtual keyboard is visible. It's just that the automation API can't find the buttons or something...

As another way forward, could I add some debug logging in the JavaScript code to run on the bots?

Patch-set: 2
Reviewer: Gerrit User 1197444 <1197444@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>