Update patch set 3
Patch Set 3: Code-Review+2

nice!

part of me wonders if this should be a skippable pre-upload warning or something similar, though, since i'm not completely sure that e.g. every printer test should depend on "cups", and since if people get this wrong, we'll realize it later anyway because the test will fail when it's run on a DUT that doesn't support it.

putting it another way: if every test in a given category is expected to have some set of dependencies, could those dependencies be specified by the category so they don't need to be copy-and-pasted in each test?

but this seems fine to me for now (modulo package naming suggestion on the other change). :-)

Patch-set: 3
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2