DO NOT REVIEW: tast-tests: add platform/memory_pressure.go
(uploading for shared testing)

This is a rewrite of the platform_MemoryPressure autotest.
There are slight changes in functionality.
Most importantly, it uses Web Page Replay to serve all requests.

This test reproduces a streamlined user behavior, with the intent
of filling up RAM quickly in a somewhat realistic way.  It opens
a bunch of tabs from a list, going over the list multiple times
as needed.  After opening each tab, it cycles through a small
number of recently-used tabs.  After the first tab discard
(or OOM kill) it collects some data, then it quiesces and
collects some more data.

BUG=None
TEST=None

Change-Id: I6e778b4173ad4b43eeb2abff093cc462482f7c6d